Configs: Initial commit for common Dsp configuration
Add a common configuration files used by samples
and other dsp baseimages. The files includes the IPC
common configuration satisfying the OMAP4 dsp platform
target.

Configure GPT5 to be used as system tick on dsp core.
The configuration is done manually to enable the wugen bit and
set the idlemode as smart-idle with wake-up. This way the
GPT5 could be able to generate the interupt even if is the
only module in its domain (and ICLK has been gated)

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>