IPC: Rearrange mboxId-procId relation
Mpu side uses two mailboxes for a duplex communication
between Mpu and remoteproc. The distribution is as follows:

mailbox_1:
   mpu --> rproc1 : mbox 0
   mpu <-- rproc1 : mbox 1

mailbox_2
   mpu <-- rproc2 : mbox 2
   mpu --> rproc2 : mbox 3

rproc1 is used for abstracting both M3 cores in one entity.
mbox0/mbox1 are properly reserved for Rx/Tx in Core0.
However, mbox 2 is used for Rx in Core1 due to the rerouting
logic. This is not correct since mbox2 should be reserved for
a different remoteproc like dsp.
This patch assigns mbox 4 for Rx in Core1 to avoid conflicts
when both remoteprocs (M3's and dsp) are enabled together.
Also mbox nomenclature has been modified to reflect the
transmision direction.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>

Conflicts:

	src/ti/ipc/rpmsg/InterruptM3.c